FIX https redirections
Some gems are reported to be hosted on `http://github.com` rather than
the `https` version of github. This leads to a redirection which is not
handled by `open-uri`.
Since updating `scheme` after uri creation does not seem to work
anymore, check url before converting is to an `URI`.

Details
* FIX url conversion to `https`