ADD auto-commit option
When passed `--commit` option, gem_update will automatically create a
commit message using the generated diff.

User will be given a chance to edit message before committing. Note that
to allow that, git commit's `-t` option has been used. While it's cool
because it allows both to provide a message and to edit it, it has a
limitation: unmodified message will result in cancelled commit.

To fix that, `--allow-empty-message` has been added. It has problems of
its own: user can't cancel commit manually. It's still possible to
reverse commit if really that was not wanted (after all, user
explicitely asked for it).

It won't be any trouble if there was nothing to update, in case you
wonder: git will see there's nothing to commit and won't even open the
editor.