FIX rubygems returning blank values
We look for an url to parse in response from rubygems.
Ideally, if there is no url, response should not even define it,
but for some gems it will return an empty string,
leading to `open-uri` crashing when attempting to open a blank url.

We should not only check if something is returned, but also that it is
not an empty string.

Related #43

Details
* FIX considering `''` as an url
* UPDATE specs