FIX using other source with rubygems
When using an other source (a geminabox install in my case) along with
rubygems, gem_updater will crash when trying to find the changelog

```
uri/common.rb:715:in `URI': bad argument (expected URI object or URI string) (ArgumentError)
        from gem_updater-0.4.5/lib/gem_updater/source_page_parser.rb:44:in `correct_uri'
        from gem_updater-0.4.5/lib/gem_updater/source_page_parser.rb:13:in `initialize'
        from gem_updater-0.4.5/lib/gem_updater.rb:51:in `new'
        from gem_updater-0.4.5/lib/gem_updater.rb:51:in `block (2 levels) in fill_changelogs'
```

With this patch, it will simply ignore an unusable url. Don't ask me
why, but the url passed to `SourcePageParser#correct_uri` for a
geminabox gem was an `IO` instead of being a `String`.